--- apps/nextjs/src/app/dashboard/@sidebar/@sidebar-header/page.tsx
@@ -1,13 +1,0 @@
-import { SidebarHeader } from "@battle-stadium/ui/sidebar";
-
-import { MainAppLinksMenu } from "./links-menu";
-
-const SIDEBAR_DEFAULT_CLASSNAMES = "m-4 rounded-xl bg-neutral-900/50";
-
-export default function SidebarHeaderSlot() {
-  return (
-    <SidebarHeader className={SIDEBAR_DEFAULT_CLASSNAMES}>
-      <MainAppLinksMenu />
-    </SidebarHeader>
-  );
-}
