+++ apps/nextjs/src/app/dashboard/@sidebar/@sidebarfooter/default.tsx
@@ -1,0 +1,5 @@
+import { NavUserComponent } from "../_components/nav-user";
+
+export default function SidebarFooterSlot() {
+  return <NavUserComponent />;
+}
