+++ apps/nextjs/src/app/dashboard/@sidebar/@sidebarheader/page.tsx
@@ -1,0 +1,5 @@
+import { MainAppLinksMenu } from "./links-menu";
+
+export default function SidebarHeaderSlot() {
+  return <MainAppLinksMenu />;
+}
