+++ apps/nextjs/src/app/dashboard/@sidebar/layout.tsx
@@ -1,17 +1,31 @@
-import { Sidebar } from "@battle-stadium/ui/sidebar";
+import { cn } from "@battle-stadium/ui";
+import {
+  Sidebar,
+  SidebarFooter,
+  SidebarHeader,
+} from "@battle-stadium/ui/sidebar";
 
 interface SideBarSlots {
-  "sidebar-header": React.ReactNode;
-  "sidebar-footer": React.ReactNode;
-  children?: never;
+  sidebarheader: React.ReactNode;
+  sidebarfooter: React.ReactNode;
+  children?: React.ReactNode;
 }
-export default function SideBar(props: SideBarSlots) {
+
+const SIDEBAR_DEFAULT_CLASSNAMES = "m-4 rounded-xl bg-neutral-900/50";
+export default function SideBar({
+  children,
+  sidebarfooter,
+  sidebarheader,
+}: SideBarSlots) {
   return (
     <Sidebar variant="sidebar" collapsible="icon">
-      {props["sidebar-header"]}
-
-      {props.children}
-      {props["sidebar-footer"]}
+      <SidebarHeader className={SIDEBAR_DEFAULT_CLASSNAMES}>
+        {sidebarheader}
+      </SidebarHeader>
+      {children}
+      <SidebarFooter className={cn(SIDEBAR_DEFAULT_CLASSNAMES, "mt-0")}>
+        {sidebarfooter}
+      </SidebarFooter>
     </Sidebar>
   );
 }
