+++ cspell.config.yaml
@@ -60,6 +60,8 @@ words:
   - pokepaste
   - randompic
   - scrollback
+  - sidebarfooter
+  - sidebarheader
   - sinnoh
   - slotnumber
   - sqlstring
